refactor: Rename NetPromoter folder to ExpandedModal and extract questionList into its own file
