+++ src/components/common/QuestionnaireLabel/ExpandedModal/NetPromoter.module.css
