+++ src/components/common/QuestionnaireLabel/ExpandedModal/ScoreRange.js
