+++ src/components/common/QuestionnaireLabel/ExpandedModal/laborImage-full.png
