+++ src/components/common/QuestionnaireLabel/ExpandedModal/laborImage.png
